+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Create and update operation using bloom filters.
+- Text match querying using bloom filters.
 
 ### Fixed